Extract splinefont_enums.h
