+++ fontforgeexe/shapers/harfbuzz.cpp
@@ -28,6 +28,7 @@
 extern "C" {
 #include "gfile.h"
 #include "splinechar.h"
+#include "splinefont_enums.h"
 #include "utype.h"
 }
 
@@ -37,10 +38,9 @@ HarfBuzzShaper::HarfBuzzShaper(std::shared_ptr<ShaperContext> context)
     : context_(context) {
     FILE* ttf_file = GFileTmpfile();
 
-    _WriteTTFFont(
-        ttf_file, context_->sf, 13 /* ff_ttf */, NULL, 1 /*bf_ttf*/,
-        32 | (1 << 29) /* ttf_flag_otmode | ttf_flag_oldkernmappedonly */,
-        context_->get_enc_map(context_->sf), 1 /*ly_fore*/);
+    _WriteTTFFont(ttf_file, context_->sf, ff_ttf, NULL, bf_ttf,
+                  ttf_flag_otmode | ttf_flag_oldkernmappedonly,
+                  context_->get_enc_map(context_->sf), ly_fore);
 
     // Calculate file length
     fseek(ttf_file, 0L, SEEK_END);
